Fix for issue #14 - Use GitHub API for each project
Preview: http://imgur.com/a/pT7dy .

The fix is based on some previous work done by @zupo in the
automatic_project branch.  He modified the index and the projects templates
to list the most active projects based on data stored in projects.json.

I wrote a simple python script that uses the github api to add/update some
stats (such as lastupdated, number of forks, number of issues to be fixed
etc.) to that json file.  Then, I updated the aformentioned templates to
display that data.  Finally I made some css tweaks.

The update script can be run from cron. A sample crontab file is attached.

Note: it seems that the api is currently a bit bugged: istead of
watchers_count, it shows stargazers_count.  Until that's fixed, the watchers
number is commented out.

Hope everything else is in due order.